1.5.0
Voucher compat with Cryptid shenanigans, overhaul of Smeltery and Prospect's material system to be a bit more like OreDictionary of modded Minecraft. MoreFluff compat for Gemstone and Brass for minerals. +2 Partners when Partners is installed. Overhauled the Config tab, moved logging options to its' own tab in the mod menu and added DEBUG messages whenever most hooks by this mod are executed.